Fixed merge
Still need to make the pages implement Author repository
