+++ src/Chirp.Core/IAuthorRepository.cs
@@ -6,5 +6,6 @@ public interface IAuthorRepository
     public Task<AuthorDTO> GetAuthorByEmail(string email);
     public Task<AuthorDTO> GetAuthorByName(string name);
     public Task<AuthorDTO> GetAuthor(string name, string email);
+    public Task<List<AuthorDTO>> GetValidUsernames(List<string> mentions);
     public void AnonymizeUser(string name, string email);
 }
